Update Roadmap.md
Added descriptions for the SQL links, replacing the !TODO! with what I think so far might be good description. Since they're all video, I also added the average duration for each.
And for entry 4 in sql, I changed 'links to the exercises' to 'Link to the video', as this is a more concept base one as far as I can see, though the principles described there (Design and E-R diagrams) will certainly be more than useful in a real business like environment.

May I ask someone to review the changes? I've put them in a new pull request to have it double check before merging.