Update Roadmap.md
I added links in section 4.0 and 5.0 ... but I do have some issues with the html 'markers' ... I think I forgot something to 'hide' the link into a zone 'link to the course', like was done here before (for clarity purposes). 

And I messed up something somewhere as the single link that I have put on Deep Learning ended up at the end of the roadmap for some reason. :O I assume I must have forgotten a marker to indicate the end of a section or something, which might be why it is messed up right now.

Can I ask someone who's a bit better to correct and, after having done so, to tell me and everyone else where I messed up? It would be useful and will prevent more avoidable corrections in the future and, in general, would make life easier for everyone here I believe :)